requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via bs4
brotli==1.1.0
    # via yt-dlp
bs4==0.0.2
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
    #   yt-dlp
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   langsmith
    #   openai
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.16
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.16
    # via -r requirements.in
langchain-core==0.2.38
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.2.4
    # via langchain
langsmith==0.1.116
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.22.0
    # via dataclasses-json
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mutagen==1.47.0
    # via yt-dlp
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.43.0
    # via -r requirements.in
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   langchain-core
    #   marshmallow
pycryptodomex==3.20.0
    # via yt-dlp
pydantic==2.9.0
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.23.2
    # via pydantic
pydub==0.25.1
    # via -r requirements.in
pypdf==4.3.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   yt-dlp
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.34
    # via
    #   langchain
    #   langchain-community
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pydantic
urllib3==2.2.2
    # via
    #   requests
    #   yt-dlp
websockets==13.0.1
    # via yt-dlp
yarl==1.10.0
    # via aiohttp
yt-dlp==2024.8.6
    # via -r requirements.in
